requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
supabase==2.3.4
PyJWT==2.10.1
gunicorn==21.2.0
pytz==2025.2
